Move cut segnments into python #10
