--- bin/cut-segments
@@ -1,21 +1,0 @@
-#!/bin/bash
-
-set -e
-
-echo "bin/cut-segments"
-
-main(){
-    input="video/2024-01-13 08.38.31.mov"
-    while IFS=' ' read -r start duration; do
-        ffmpeg -nostdin -i "$input" -ss "$start" -t "$duration"  -c:v libx264 -c:a aac -metadata:s:v rotate="0" "output_${start//[:]/-}_to_${duration//[:]/-}.mp4"
-    done < "output/segments.txt"
-
-}
-
-main
-
-#!/bin/bash
-
-# Input file
-
-# Read segments from file
